+++ my_pipeline/__init__.py
@@ -14,3 +14,4 @@
 
 
 from .pipeline import *
+from .transforms import *
